Extract Import types to Dhall.Syntax.Import
Context: #2313
