+++ dhall/src/Dhall/Binary.hs
@@ -32,21 +32,23 @@ import Dhall.Syntax
     , Chunks (..)
     , Const (..)
     , DhallDouble (..)
-    , Directory (..)
     , Expr (..)
+    , FunctionBinding (..)
+    , MultiLet (..)
+    , PreferAnnotation (..)
+    , RecordField (..)
+    , Var (..)
+    )
+import Dhall.Syntax.Import
+    ( Directory (..)
     , File (..)
     , FilePrefix (..)
-    , FunctionBinding (..)
     , Import (..)
     , ImportHashed (..)
     , ImportMode (..)
     , ImportType (..)
-    , MultiLet (..)
-    , PreferAnnotation (..)
-    , RecordField (..)
     , Scheme (..)
     , URL (..)
-    , Var (..)
     )
 
 import Data.Foldable (toList)
