+++ dhall/src/Dhall/Schemas.hs
@@ -20,11 +20,11 @@ import Control.Exception   (Exception)
 import Data.Maybe          (fromMaybe)
 import Data.Text           (Text)
 import Data.Void           (Void)
+import Dhall.Core          (Expr (..), Import, Var (..))
 import Dhall.Crypto        (SHA256Digest)
 import Dhall.Map           (Map)
 import Dhall.Pretty        (CharacterSet (..), detectCharacterSet)
 import Dhall.Src           (Src)
-import Dhall.Syntax        (Expr (..), Import, Var (..))
 import Dhall.Util
     ( Censor (..)
     , Header (..)
