+++ dhall/src/Dhall/Syntax.hs
@@ -17,6 +17,8 @@
 
 {-| This module contains the core syntax types and optics for them.
 
+The import types are excluded and can be found in "Dhall.Syntax.Import".
+
 'reservedIdentifiers', 'denote' and friends are included because they are
 involved in a dependency circle with "Dhall.Pretty.Internal".
 -}
@@ -58,18 +60,6 @@ module Dhall.Syntax (
     , renote
     , shallowDenote
 
-    -- * 'Import'
-    , Directory(..)
-    , File(..)
-    , FilePrefix(..)
-    , Import(..)
-    , ImportHashed(..)
-    , ImportMode(..)
-    , ImportType(..)
-    , URL(..)
-    , Scheme(..)
-    , pathCharacter
-
     -- * Reserved identifiers
     , reservedIdentifiers
     , reservedKeywords
@@ -109,7 +99,7 @@ import                GHC.Generics               (Generic)
 import                Instances.TH.Lift          ()
 import                Language.Haskell.TH.Syntax (Lift)
 import                Numeric.Natural            (Natural)
-import                Prettyprinter              (Doc, Pretty)
+import                Prettyprinter              (Pretty)
 import                Unsafe.Coerce              (unsafeCoerce)
 
 import qualified Control.Monad
@@ -118,10 +108,8 @@ import qualified Data.HashSet
 import qualified Data.List.NonEmpty as NonEmpty
 import qualified Data.Text
 import qualified Data.Time          as Time
-import qualified Dhall.Crypto
 import qualified Dhall.Optics       as Optics
 import qualified Lens.Family        as Lens
-import qualified Network.URI        as URI
 import qualified Prettyprinter      as Pretty
 
 deriving instance Lift Time.Day
@@ -955,220 +943,6 @@ chunkExprs
 chunkExprs f (Chunks chunks final) =
   flip Chunks final <$> traverse (traverse f) chunks
 {-# INLINABLE chunkExprs #-}
-
-{-| Internal representation of a directory that stores the path components in
-    reverse order
-
-    In other words, the directory @\/foo\/bar\/baz@ is encoded as
-    @Directory { components = [ "baz", "bar", "foo" ] }@
--}
-newtype Directory = Directory { components :: [Text] }
-    deriving stock (Eq, Generic, Ord, Show)
-    deriving anyclass NFData
-
-instance Semigroup Directory where
-    Directory components₀ <> Directory components₁ =
-        Directory (components₁ <> components₀)
-
-instance Pretty Directory where
-    pretty (Directory {..}) = foldMap prettyPathComponent (reverse components)
-
-prettyPathComponent :: Text -> Doc ann
-prettyPathComponent text
-    | Data.Text.all pathCharacter text =
-        "/" <> Pretty.pretty text
-    | otherwise =
-        "/\"" <> Pretty.pretty text <> "\""
-
-{-| A `File` is a `directory` followed by one additional path component
-    representing the `file` name
--}
-data File = File
-    { directory :: Directory
-    , file      :: Text
-    } deriving (Eq, Generic, Ord, Show, NFData)
-
-instance Pretty File where
-    pretty (File {..}) =
-            Pretty.pretty directory
-        <>  prettyPathComponent file
-
-instance Semigroup File where
-    File directory₀ _ <> File directory₁ file =
-        File (directory₀ <> directory₁) file
-
--- | The beginning of a file path which anchors subsequent path components
-data FilePrefix
-    = Absolute
-    -- ^ Absolute path
-    | Here
-    -- ^ Path relative to @.@
-    | Parent
-    -- ^ Path relative to @..@
-    | Home
-    -- ^ Path relative to @~@
-    deriving (Eq, Generic, Ord, Show, NFData)
-
-instance Pretty FilePrefix where
-    pretty Absolute = ""
-    pretty Here     = "."
-    pretty Parent   = ".."
-    pretty Home     = "~"
-
--- | The URI scheme
-data Scheme = HTTP | HTTPS deriving (Eq, Generic, Ord, Show, NFData)
-
--- | This type stores all of the components of a remote import
-data URL = URL
-    { scheme    :: Scheme
-    , authority :: Text
-    , path      :: File
-    , query     :: Maybe Text
-    , headers   :: Maybe (Expr Src Import)
-    } deriving (Eq, Generic, Ord, Show, NFData)
-
-instance Pretty URL where
-    pretty (URL {..}) =
-            schemeDoc
-        <>  "://"
-        <>  Pretty.pretty authority
-        <>  pathDoc
-        <>  queryDoc
-        <>  foldMap prettyHeaders headers
-      where
-        prettyHeaders h =
-          " using (" <> Pretty.unAnnotate (Pretty.pretty h) <> ")"
-
-        File {..} = path
-
-        Directory {..} = directory
-
-        pathDoc =
-                foldMap prettyURIComponent (reverse components)
-            <>  prettyURIComponent file
-
-        schemeDoc = case scheme of
-            HTTP  -> "http"
-            HTTPS -> "https"
-
-        queryDoc = case query of
-            Nothing -> ""
-            Just q  -> "?" <> Pretty.pretty q
-
-prettyURIComponent :: Text -> Doc ann
-prettyURIComponent text =
-        Pretty.pretty $ URI.normalizeCase $ URI.normalizeEscape $ "/" <> Data.Text.unpack text
-
--- | The type of import (i.e. local vs. remote vs. environment)
-data ImportType
-    = Local FilePrefix File
-    -- ^ Local path
-    | Remote URL
-    -- ^ URL of remote resource and optional headers stored in an import
-    | Env  Text
-    -- ^ Environment variable
-    | Missing
-    deriving (Eq, Generic, Ord, Show, NFData)
-
-parent :: File
-parent = File { directory = Directory { components = [ ".." ] }, file = "" }
-
-instance Semigroup ImportType where
-    Local prefix file₀ <> Local Here file₁ = Local prefix (file₀ <> file₁)
-
-    Remote (URL { path = path₀, ..}) <> Local Here path₁ =
-        Remote (URL { path = path₀ <> path₁, ..})
-
-    Local prefix file₀ <> Local Parent file₁ =
-        Local prefix (file₀ <> parent <> file₁)
-
-    Remote (URL { path = path₀, .. }) <> Local Parent path₁ =
-        Remote (URL { path = path₀ <> parent <> path₁, .. })
-
-    import₀ <> Remote (URL { headers = headers₀, .. }) =
-        Remote (URL { headers = headers₁, .. })
-      where
-        importHashed₀ = Import (ImportHashed Nothing import₀) Code
-
-        headers₁ = fmap (fmap (importHashed₀ <>)) headers₀
-
-    _ <> import₁ =
-        import₁
-
-instance Pretty ImportType where
-    pretty (Local prefix file) =
-        Pretty.pretty prefix <> Pretty.pretty file
-
-    pretty (Remote url) = Pretty.pretty url
-
-    pretty (Env env) = "env:" <> prettyEnvironmentVariable env
-
-    pretty Missing = "missing"
-
--- | How to interpret the import's contents (i.e. as Dhall code or raw text)
-data ImportMode = Code | RawText | Location
-  deriving (Eq, Generic, Ord, Show, NFData)
-
--- | A `ImportType` extended with an optional hash for semantic integrity checks
-data ImportHashed = ImportHashed
-    { hash       :: Maybe Dhall.Crypto.SHA256Digest
-    , importType :: ImportType
-    } deriving (Eq, Generic, Ord, Show, NFData)
-
-instance Semigroup ImportHashed where
-    ImportHashed _ importType₀ <> ImportHashed hash importType₁ =
-        ImportHashed hash (importType₀ <> importType₁)
-
-instance Pretty ImportHashed where
-    pretty (ImportHashed  Nothing p) =
-      Pretty.pretty p
-    pretty (ImportHashed (Just h) p) =
-      Pretty.group (Pretty.flatAlt long short)
-      where
-        long =
-            Pretty.align
-                (   Pretty.pretty p <> Pretty.hardline
-                <>  "  sha256:" <> Pretty.pretty (show h)
-                )
-
-        short = Pretty.pretty p <> " sha256:" <> Pretty.pretty (show h)
-
--- | Reference to an external resource
-data Import = Import
-    { importHashed :: ImportHashed
-    , importMode   :: ImportMode
-    } deriving (Eq, Generic, Ord, Show, NFData)
-
-instance Semigroup Import where
-    Import importHashed₀ _ <> Import importHashed₁ code =
-        Import (importHashed₀ <> importHashed₁) code
-
-instance Pretty Import where
-    pretty (Import {..}) = Pretty.pretty importHashed <> Pretty.pretty suffix
-      where
-        suffix :: Text
-        suffix = case importMode of
-            RawText  -> " as Text"
-            Location -> " as Location"
-            Code     -> ""
-
-{-| Returns `True` if the given `Char` is valid within an unquoted path
-    component
-
-    This is exported for reuse within the @"Dhall.Parser.Token"@ module
--}
-pathCharacter :: Char -> Bool
-pathCharacter c =
-         '\x21' == c
-    ||  ('\x24' <= c && c <= '\x27')
-    ||  ('\x2A' <= c && c <= '\x2B')
-    ||  ('\x2D' <= c && c <= '\x2E')
-    ||  ('\x30' <= c && c <= '\x3B')
-    ||  c == '\x3D'
-    ||  ('\x40' <= c && c <= '\x5A')
-    ||  ('\x5E' <= c && c <= '\x7A')
-    ||  c == '\x7C'
-    ||  c == '\x7E'
 
 -- | Remove all `Note` constructors from an `Expr` (i.e. de-`Note`)
 --
